Python/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
more-itertools==4.3.0
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.6.0
pytest==3.8.2
six==1.11.0
